login_service/requirements.txt
flask
flask_prometheus_metrics
Flask-Session
prometheus_flask_exporter
pyjwt
pymongo
pytest
requests